+++ changelog.md
@@ -13,6 +13,8 @@ those changes:
   requests per day to 25000 requests per day.
 * The Developer plan is renamed to Plus and the limit is changed from 500000
   requests per day to 250000 requests per day.
+* The Advanced plan is renamed to Business and the limit is changed from
+  5000000 requests per day to 2000000 requests per day.
 
 simdjson, the library used for JSON parsing, has been updated from version
 3.11.1 to version 3.11.6.